feat: 랜딩 타블렛, 모바일 카드 스켈레톤 적용
